% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q1.R
\name{Q.1}
\alias{Q.1}
\title{First-lag transition matrix}
\usage{
Q.1(q)
}
\arguments{
\item{q}{the observed marginal proportions}
}
\description{
This function computes the first-lag transition matrix from
the observed marginal proportions q
}
