% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2.R
\name{Q.2}
\alias{Q.2}
\title{Second-lag transition matrix}
\usage{
Q.2(q)
}
\arguments{
\item{q}{the observed marginal proportions}
}
\description{
This function computes the second-lag transition matrix from
the observed marginal proportions q
}
