% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnq.R
\name{v.n.q}
\alias{v.n.q}
\title{Asymptotic variance of the Multinomial model}
\usage{
v.n.q(n, m, q)
}
\arguments{
\item{n}{The number of possible bins (patterns)}

\item{m}{The number of observed patterns}

\item{q}{The observed proportions (in no specific order)

#' X <- rnorm(1000) # a time series
q1 <- pdc::codebook(X, m=3, t=1) # computes the ordinal patterns
v.n.q() # COMPLETAR}
}
\description{
This function computes the asymptotic variance of the Multinomial model
with proportions q. It is provided as a comparison with the asymptotic
variance of the Ordinal Patterns (that embeds the serial correlation)
}
